move IntervalMapping to dace utils, add __init__.py in dace backend
