+++ src/gtc/dace/utils.py
@@ -473,3 +473,119 @@ def iter_vertical_loop_section_sub_sdfgs(graph: SDFG) -> Iterator[SDFG]:
     for node, _ in graph.all_nodes_recursive():
         if isinstance(node, VerticalLoopLibraryNode):
             yield from (subgraph for _, subgraph in node.sections)
+
+
+class IntervalMapping:
+    def __init__(self) -> None:
+        self.interval_starts: List[oir.AxisBound] = list()
+        self.interval_ends: List[oir.AxisBound] = list()
+        self.values: List[Any] = list()
+
+    def _setitem_subset_of_existing(self, i: int, key: oir.Interval, value: Any) -> None:
+        start = self.interval_starts[i]
+        end = self.interval_ends[i]
+        if self.values[i] is not value:
+            idx = i
+            if key.start != start:
+                self.interval_ends[i] = key.start
+                self.interval_starts.insert(i + 1, key.start)
+                self.interval_ends.insert(i + 1, key.end)
+                self.values.insert(i + 1, value)
+                idx = i + 1
+            if key.end != end:
+                self.interval_starts.insert(idx + 1, key.end)
+                self.interval_ends.insert(idx + 1, end)
+                self.values.insert(idx + 1, self.values[i])
+                self.interval_ends[idx] = key.end
+                self.values[idx] = value
+
+    def _setitem_partial_overlap(self, i: int, key: oir.Interval, value: Any) -> None:
+        start = self.interval_starts[i]
+        if key.start < start:
+            if self.values[i] is value:
+                self.interval_starts[i] = key.start
+            else:
+                self.interval_starts[i] = key.end
+                self.interval_starts.insert(i, key.start)
+                self.interval_ends.insert(i, key.end)
+                self.values.insert(i, value)
+        else:  # key.end > end
+            if self.values[i] is value:
+                self.interval_ends[i] = key.end
+                nextidx = i + 1
+            else:
+                self.interval_ends[i] = key.start
+                self.interval_starts.insert(i + 1, key.start)
+                self.interval_ends.insert(i + 1, key.end)
+                self.values.insert(i + 1, value)
+                nextidx = i + 2
+            if nextidx < len(self.interval_starts) and (
+                key.intersects(
+                    oir.Interval(
+                        start=self.interval_starts[nextidx], end=self.interval_ends[nextidx]
+                    )
+                )
+                or self.interval_starts[nextidx] == key.end
+            ):
+                if self.values[nextidx] is value:
+                    self.interval_ends[nextidx - 1] = self.interval_ends[nextidx]
+                    del self.interval_starts[nextidx]
+                    del self.interval_ends[nextidx]
+                    del self.values[nextidx]
+                else:
+                    self.interval_starts[nextidx] = key.end
+
+    def __setitem__(self, key: oir.Interval, value: Any) -> None:
+        if not isinstance(key, oir.Interval):
+            raise TypeError("Only OIR intervals supported for method add of IntervalSet.")
+
+        delete = list()
+        for i, (start, end) in enumerate(zip(self.interval_starts, self.interval_ends)):
+            if key.covers(oir.Interval(start=start, end=end)):
+                delete.append(i)
+
+        for i in reversed(delete):  # so indices keep validity while deleting
+            del self.interval_starts[i]
+            del self.interval_ends[i]
+            del self.values[i]
+
+        if len(self.interval_starts) == 0:
+            self.interval_starts.append(key.start)
+            self.interval_ends.append(key.end)
+            self.values.append(value)
+            return
+
+        for i, (start, end) in enumerate(zip(self.interval_starts, self.interval_ends)):
+            if oir.Interval(start=start, end=end).covers(key):
+                self._setitem_subset_of_existing(i, key, value)
+                return
+
+        for i, (start, end) in enumerate(zip(self.interval_starts, self.interval_ends)):
+            if (
+                key.intersects(oir.Interval(start=start, end=end))
+                or start == key.end
+                or end == key.start
+            ):
+                self._setitem_partial_overlap(i, key, value)
+                return
+
+        for i, start in enumerate(self.interval_starts):
+            if start > key.start:
+                self.interval_starts.insert(i, key.start)
+                self.interval_ends.insert(i, key.end)
+                self.values.insert(i, value)
+                return
+        self.interval_starts.append(key.start)
+        self.interval_ends.append(key.end)
+        self.values.append(value)
+        return
+
+    def __getitem__(self, key: oir.Interval) -> List[Any]:
+        if not isinstance(key, oir.Interval):
+            raise TypeError("Only OIR intervals supported for keys of IntervalMapping.")
+
+        res = []
+        for start, end, value in zip(self.interval_starts, self.interval_ends, self.values):
+            if key.intersects(oir.Interval(start=start, end=end)):
+                res.append(value)
+        return res
